+++ .eslintrc.js
@@ -25,6 +25,7 @@ module.exports = {
     "unicorn/no-array-for-each": "off",
     "unicorn/prefer-node-protocol": "off",
     "unicorn/prefer-module": "off",
+    "unicorn/consistent-destructuring": "off",
     "unicorn/filename-case": [
       "error",
       {
@@ -55,6 +56,7 @@ module.exports = {
         "unicorn/no-nested-ternary": "off",
         "unicorn/no-null": "off",
         "unicorn/prefer-spread": "off",
+        "unicorn/consistent-destructuring": "off",
         "unicorn/no-useless-undefined": [
           "error",
           {
@@ -81,6 +83,7 @@ module.exports = {
             checkFilenames: false,
           },
         ],
+        "unicorn/no-array-reduce": "off",
       },
     },
   ],